[python-frontend]  move Python type-annotation checks into helper class
